requirements.txt
flask==3.0.2
opensearch-py==2.4.2
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.0.1
Werkzeug==3.0.1 